05_agent_memory/pyproject.toml
[project]
name = "travel-agent-memory"
version = "0.1.0"
description = "Travel agent with AgentCore memory demonstration"
requires-python = ">=3.10"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "rich",
]

[tool.hatch.build.targets.wheel]
packages = []